\name{rdcor-package}
\alias{rdcor-package}
\docType{package}
\title{
Rank Distance Correlation Coefficient
}
\description{
The rank distance correlation of Shi H., Drton M. and Han F. (2022) is computed. Included also is a function to perform permutation based testing.
}
\details{
\tabular{lll}{
Package: \tab rdcor\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2026-03-25\cr
License: \tab GPL-2\cr
}
}
\author{
Michail Tsagris \email{mtsagris@uoc.gr}.
}

\section{Maintainers}{
Michail Tsagris \email{mtsagris@uoc.gr}.
}

\references{
Shi H., Drton M. and Han F. (2022). Distribution-free consistent independence tests via center-outward ranks and signs. Journal of the American Statistical Association, 117(537): 395--410.

Zhang Q. (2025). On the connections between Chatterjee's correlation and rank distance correlation.
Journal of Nonparametric Statistics, 1--18.
}



